Move redirection to static.js
