+++ lib/assets/javascripts/cartodb/helpers/parse-domains.js
@@ -1,4 +1,4 @@
-const URL = require('url-parse');
+var URL = require('url-parse');
 
 /**
  * @param {string} url
