+++ lib/assets/javascripts/cartodb/static.js
@@ -1,6 +1,7 @@
 var CartoNode = require('../../../../vendor/assets/javascripts/carto-node/carto-node.js');
 var PACKAGE = require('../../../../package.json');
 var UrlHelper = require('./helpers/url');
+var redirector = require('./helpers/redirector');
 
 var version = PACKAGE.version;
 var AssetsVersionHelper = require('./helpers/assets_version');
@@ -21,6 +22,11 @@ window.CartoConfig = window.CartoConfig || {};
   var addSpinner = function () {
     var spinnerEl = document.querySelector('.spinner');
     spinnerEl.parentNode.classList.add('is-hidden');
+  };
+
+  var redirectUsers = function (userData) {
+    redirector.redirectOrgUsers(userData.organization, userData.username, window.location);
+    redirector.redirectToLoginIfAnotherIsLogged(userData.username, window.location);
   };
 
   var addAssets = function () {
@@ -79,6 +85,10 @@ window.CartoConfig = window.CartoConfig || {};
 
       if (!data.user_data && visualizationError !== VISUALIZATION_ERROR) {
         return redirectToLogin();
+      }
+
+      if (data.user_data) {
+        redirectUsers(data.user_data);
       }
 
       window.CartoConfig.data = data;
