+++ lib/assets/javascripts/dashboard/account.js
@@ -8,7 +8,6 @@ const ConfigModel = require('dashboard/data/config-model');
 const UserModel = require('dashboard/data/user-model');
 const OrganizationModel = require('dashboard/data/organization-model');
 const AssetsVersionHelper = require('dashboard/helpers/assets-version');
-const redirector = require('dashboard/helpers/redirector');
 
 const Locale = require('locale/index');
 
@@ -33,8 +32,6 @@ const InitAccount = function () {
     const ASSETS_VERSION = AssetsVersionHelper.getAssetsVersion(PACKAGE.version);
     const dashboard_notifications = data.dashboard_notifications;
     const userData = data.user_data || {};
-    redirector.redirectOrgUsers(userData.organization, userData.username, 'account', window.location);
-    redirector.redirectToLoginIfAnotherIsLogged(userData.username, window.location);
 
     const configModel = new ConfigModel(
       _.extend(
