+++ lib/assets/javascripts/dashboard/dashboard.js
@@ -2,7 +2,6 @@ const _ = require('underscore');
 
 const Locale = require('locale/index');
 const Polyglot = require('node-polyglot');
-const redirector = require('dashboard/helpers/redirector');
 
 const ACTIVE_LOCALE = 'en';
 
@@ -52,9 +51,6 @@ function dataLoaded (data) {
   const dashboard_notifications = data.dashboard_notifications;
   const isJustLoggedIn = data.isJustLoggedIn;
   const isFirstTimeViewingDashboard = data.isFirstTimeViewingDashboard;
-
-  redirector.redirectOrgUsers(userData.organization, userData.username, 'dashboard', window.location);
-  redirector.redirectToLoginIfAnotherIsLogged(userData.username, window.location);
 
   const configModel = new ConfigModel(
     _.extend(
