+++ lib/assets/javascripts/dashboard/profile.js
@@ -9,7 +9,6 @@ const ConfigModel = require('dashboard/data/config-model');
 const UserModel = require('dashboard/data/user-model');
 const ModalsServiceModel = require('builder/components/modals/modals-service-model');
 const AssetsVersionHelper = require('dashboard/helpers/assets-version');
-const redirector = require('dashboard/helpers/redirector');
 
 const Locale = require('locale/index');
 const Polyglot = require('node-polyglot');
@@ -34,9 +33,6 @@ window.CartoConfig = window.CartoConfig || {};
 
 function dataLoaded (data) {
   const {user_data, dashboard_notifications} = data;
-
-  redirector.redirectOrgUsers(user_data.organization, user_data.username, 'profile', window.location);
-  redirector.redirectToLoginIfAnotherIsLogged(user_data.username, window.location);
 
   const configModel = new ConfigModel(
     _.extend(
